@@ -1,121 +1,351 @@
 ﻿using Assimp;
 using Assimp.Configs;
-using Engine.Components;
+using Engine.Animations;
+using Engine.Extensions;
 using Engine.GLObjects.Textures;
-using Engine.Physics;
 using Engine.Rendering.Enums;
-using Engine.Rendering.GameObjects;
 using Engine.Structs;
 using OpenTK.Graphics.OpenGL4;
 using OpenTK.Mathematics;
 using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using Mesh = Engine.Rendering.GameObjects.Mesh;
 
 namespace Engine.GameObjects
 {
-    public class Model
+    /// <summary>
+    /// Класс модели состоит из множества мешей 
+    /// </summary>
+    public class Model : IDisposable
     {
+        /// <summary>
+        /// Словарь уже существующих текстур, где ключ - путь, значение - модель
+        /// </summary>
         private static readonly Dictionary<string, Model> _models = new();
+
+        /// <summary>
+        /// Список загруженных текстур
+        /// </summary>
         private static readonly List<Texture> _loadedTextures = new();
+
         private readonly string directory;
         private readonly string _path;
 
-        public List<BoundingRegion> BoundingRegions;
-        public List<RigidBody> RigidBodies;
+        /// <summary>
+        /// Список мешей
+        /// </summary>
         public List<Mesh> Meshes = new();
+
+        /// <summary>
+        /// Словарь костей, для анимаций
+        /// </summary>
+        private readonly Dictionary<string, BoneInfo> _boneInfoMap = new Dictionary<string, BoneInfo>();
+
+        /// <summary>
+        /// Кол-во костей
+        /// </summary>
+        private int _boneCounter = 0;
+
+        /// <summary>
+        /// Максимальный вес костей
+        /// </summary>
+        public const int MAX_BONE_WEIGHTS = 100;
+
+        /// <summary>
+        /// Геттер получения словаря костей
+        /// </summary>
+        /// <returns></returns>
+        public Dictionary<string, BoneInfo> GetBoneInfoMap() => _boneInfoMap;
+
+        /// <summary>
+        /// Геттер получения кол-во костей
+        /// </summary>
+        /// <returns></returns>
+        public int GetBoneCount() => _boneCounter;
+
+        /// <summary>
+        /// Класс отвечающий за воспроизведение анимаций
+        /// </summary>
+        public Animator Animator;
+
+        /// <summary>
+        /// Список анимаций, 
+        /// TODO: стоит возможно сделать статичным и загружать туда все анимации и воспроизводить из <see cref="Animator"/>
+        /// </summary>
+        public Dictionary<string, Animations.Animation> Animations;
+
         #region Constructor
+        /// <summary>
+        /// Конструктор модели по пути
+        /// </summary>
+        /// <param name="path">Путь, где находится модель</param>
         public Model(string path)
         {
             directory = path.Substring(0, path.LastIndexOf('/'));
             _path = path;
-            LoadModel();
+            // Если модель не существует в словаре моделей, то загружаем её
+            if (!CheckModels(_path))
+            {
+                LoadModel();
+                //GetOrSetThread();
+            }
         }
-        public Model(Mesh mesh)
-        {
-            Meshes.Add(mesh);
-            _models.Add(Guid.NewGuid().ToString(),this);
 
+        /// <summary>
+        /// Создание модели из 1 меша
+        /// </summary>
+        /// <param name="mesh">Сам меш</param>
+        /// <param name="name">Название модели, которое пойдет как путь модели</param>
+        public Model(Mesh mesh, string name)
+        {
+            if (!CheckModels(name))
+            {
+                Meshes.Add(mesh);
+
+                _path = name;
+            }
+        }
+
+        /// <summary>
+        /// Конструктор модели из 1 меша и списка текстур по имени
+        /// </summary>
+        /// <param name="mesh">Сам меш</param>
+        /// <param name="textures">Список текстур</param>
+        /// <param name="name">Наименование модели, пойдет как путь модели</param>
+        public Model(Mesh mesh, List<Texture> textures, string name)
+            : this(mesh, name)
+        {
+            if (!CheckModels(name))
+            {
+                Meshes.Add(mesh);
+
+                _path = name;
+                mesh.Textures.AddRange(textures);
+            }
+        }
+
+        private void GetOrSetThread()
+        {
+            /*            if (ImportingTask != null)
+                        {
+                            ImportingQueue.Enqueue(LoadModel);
+                        }
+                        else
+                        {
+                            ImportingQueue.Enqueue(LoadModel);
+                            ImportingTask = new Task(() =>
+                            {
+                                while (!_requestTermination.WaitOne(0))
+                                {
+                                    if (ImportingQueue.Count > 0)
+                                    {
+                                        var action = ImportingQueue.Dequeue();
+                                        action.Invoke();
+                                    }
+                                }
+                            });
+                            ImportingTask.Start();
+                        }*/
+        }
+
+        /// <summary>
+        /// Возращает true, если модель уже была загружена
+        /// </summary>
+        /// <param name="path">путь, по которому ищем в словаре моделей </param>
+        /// <returns></returns>
+        private bool CheckModels(string path)
+        {
+            if (path == null)
+            {
+                return false;
+            }
+            if (_models.TryGetValue(path, out var m))
+            {
+                Meshes.AddRange(m.Meshes);
+                return true;
+            }
+            return false;
         }
         #endregion
 
-        public void Render(Shader shader, RenderFlags flags, TextureTarget target = TextureTarget.Texture2D)
+        /// <summary>
+        /// Функция отрисовки модели
+        /// </summary>
+        /// <param name="shader">Шейдер, куда отправляем модель</param>
+        /// <param name="flags">Флаги отрисовки модели</param>
+        /// <param name="target">Куда накладываем текстуры</param>
+        /// <param name="type">Тип отрсиовки модели</param>
+        public void Render(Shader shader, RenderFlags flags, TextureTarget target = TextureTarget.Texture2D, OpenTK.Graphics.OpenGL4.PrimitiveType type = OpenTK.Graphics.OpenGL4.PrimitiveType.Triangles)
         {
-            Meshes.ForEach(s=>s.Render(shader,flags, target));
+            if (flags.HasFlag(RenderFlags.Animation))
+            {
+                if (Animator is not null && Animator.HasAnimation)
+                {
+                    Animator.Render(shader);
+                }
+            }
+            if (flags.HasFlag(RenderFlags.Mesh))
+            {
+                foreach (var item in Meshes)
+                {
+                    item.Render(shader, flags, target, type);
+                }
+            }
+
         }
+
         #region Private Methods
 
+        /// <summary>
+        /// Устанавливаем на каждую вершину данные по костям по умолчанию
+        /// </summary>
+        /// <param name="v">Вершину, которую будем изменять</param>
+        /// <returns></returns>
+        private static Vertex SetVertexBoneDataToDefault(Vertex v)
+        {
+            for (int i = 0; i < Vertex.MAX_BONE_INFLUENCE; i++)
+            {
+                v.BoneIDs[i] = -1;
+                v.Weights[i] = 0.0f;
+            }
+            return v;
+        }
+
+        /// <summary>
+        /// Устанавливаем веса для костей для определенной вершины
+        /// </summary>
+        /// <param name="v">Вершина</param>
+        /// <param name="boneID">Индекс кости</param>
+        /// <param name="weight">Вес кости</param>
+        /// <returns>Изменненная вершина</returns>
+        private static Vertex SetVertexBoneData(Vertex v, int boneID, float weight)
+        {
+            for (int i = 0; i < Vertex.MAX_BONE_INFLUENCE; i++)
+            {
+                if (v.BoneIDs[i] < 0)
+                {
+                    v.Weights[i] = weight;
+                    v.BoneIDs[i] = boneID;
+                    return v;
+                }
+            }
+            return v;
+        }
+
+        /// <summary>
+        /// Извлекаем веса вершины и устанавливаем в соответствии с словарем костей
+        /// </summary>
+        /// <param name="vertices">список вершин</param>
+        /// <param name="mesh">меш из загружаемой модели</param>
+        private void ExtractBoneWeightForVertices(List<Vertex> vertices, Assimp.Mesh mesh)
+        {
+            for (int boneIndex = 0; boneIndex < mesh.BoneCount; boneIndex++)
+            {
+                string boneName = mesh.Bones[boneIndex].Name;
+                int boneID;
+                if (!_boneInfoMap.ContainsKey(boneName))
+                {
+                    BoneInfo newBoneInfo;
+
+                    newBoneInfo.ID = _boneCounter;
+                    newBoneInfo.offset = mesh.Bones[boneIndex].OffsetMatrix.GetMatrix4FromAssimpMatrix();
+                    boneID = _boneCounter;
+                    _boneCounter++;
+                    _boneInfoMap.Add(boneName, newBoneInfo);
+                }
+                else
+                {
+                    boneID = _boneInfoMap[boneName].ID;
+                }
+                var weights = mesh.Bones[boneIndex].VertexWeights;
+                int numWeights = mesh.Bones[boneIndex].VertexWeightCount;
+                for (int weightIndex = 0; weightIndex < numWeights; ++weightIndex)
+                {
+                    int vertexId = weights[weightIndex].VertexID;
+                    float weight = weights[weightIndex].Weight;
+                    vertices[vertexId] = SetVertexBoneData(vertices[vertexId], boneID, weight);
+                }
+            }
+        }
+
+
         #region ASSIMP METHODS
-        private void ProcessNode(Assimp.Node node, Scene scene)
+        private void ProcessNode(Node node, Scene scene)
         {
             for (int i = 0; i < node.MeshCount; i++)
             {
                 Assimp.Mesh mesh = scene.Meshes[node.MeshIndices[i]];
+
                 Mesh m = ProcessMesh(mesh, scene);
                 Meshes.Add(m);
+
             }
             for (int i = 0; i < node.ChildCount; i++)
             {
                 ProcessNode(node.Children[i], scene);
             }
         }
+
+        public void Dispose()
+        {
+            Console.WriteLine($"Disposing: {_path}");
+            foreach (var item in Meshes)
+            {
+                item.Dispose();
+            }
+        }
+
         private Mesh ProcessMesh(Assimp.Mesh mesh, Scene scene)
         {
             List<Vertex> vertices = new();
-            List<int> indices = new();
+            List<uint> indices = new();
             List<Texture> textures = new();
             for (int i = 0; i < mesh.VertexCount; i++)
             {
 
-                Vertex v = new()
-                {
-                    Position = new Vector3(mesh.Vertices[i].X, mesh.Vertices[i].Y, mesh.Vertices[i].Z)
-                };
-
+                Vertex v = new(new Vector3(mesh.Vertices[i].X, mesh.Vertices[i].Y, mesh.Vertices[i].Z));
+                v = SetVertexBoneDataToDefault(v);
                 if (mesh.HasTextureCoords(0))
-                {
                     v.TexCoords = new Vector2(mesh.TextureCoordinateChannels[0][i].X, mesh.TextureCoordinateChannels[0][i].Y);
-                    if (mesh.Tangents.Count > 0)
-                    {
-                        v.Tangent = new Vector3(mesh.Tangents[i].X, mesh.Tangents[i].Y, mesh.Tangents[i].Z);
 
-                    }
-                    if (mesh.BiTangents.Count > 0)
-                    {
-                        v.Bitangent = new Vector3(mesh.BiTangents[i].X, mesh.BiTangents[i].Y, mesh.BiTangents[i].X);
-                    }
-                }
-                else
-                    v.TexCoords = new Vector2(0.0f, 0.0f);
-                if (mesh.HasNormals)
+                if (mesh.HasTangentBasis)
                 {
-                    v.Normal = new Vector3(mesh.Normals[i].X, mesh.Normals[i].Y, mesh.Normals[i].Z);
+                    v.Tangent = new Vector3(mesh.Tangents[i].X, mesh.Tangents[i].Y, mesh.Tangents[i].Z);
+                    v.Bitangent = new Vector3(mesh.BiTangents[i].X, mesh.BiTangents[i].Y, mesh.BiTangents[i].Z);
                 }
+
+                if (mesh.HasNormals)
+                    v.Normal = new Vector3(mesh.Normals[i].X, mesh.Normals[i].Y, mesh.Normals[i].Z);
 
                 vertices.Add(v);
             }
-            indices.AddRange(mesh.GetIndices());
+
+            indices.AddRange(mesh.GetIndices().Select(s => Convert.ToUInt32(s)));
             Material material = scene.Materials[mesh.MaterialIndex];
-            List<Texture> diffuseMaps = LoadMaterialTexture(material, TextureType.Diffuse, "texture_diffuse0");
+            List<Texture> diffuseMaps = LoadMaterialTexture(material, Assimp.TextureType.Diffuse, Rendering.Enums.TextureType.Diffuse);
             textures.AddRange(diffuseMaps);
 
-            List<Texture> specularMaps = LoadMaterialTexture(material, TextureType.Specular, "texture_specular0");
+            List<Texture> specularMaps = LoadMaterialTexture(material, Assimp.TextureType.Specular, Rendering.Enums.TextureType.Specular);
             textures.AddRange(specularMaps);
 
-            List<Texture> normalMaps = LoadMaterialTexture(material, TextureType.Height, "texture_normal0");
+            List<Texture> normalMaps = LoadMaterialTexture(material, Assimp.TextureType.Normals, Rendering.Enums.TextureType.Normal);
+            textures.AddRange(normalMaps);
+            List<Texture> heightMaps = LoadMaterialTexture(material, Assimp.TextureType.Height, Rendering.Enums.TextureType.Roughness);
             textures.AddRange(normalMaps);
 
-            List<Texture> heightMaps = LoadMaterialTexture(material, TextureType.Ambient, "texture_height0");
-            textures.AddRange(heightMaps);
+
             if (textures.Count == 0)
             {
-                textures.Add(Texture.LoadFromFile(Game.TEXTURES_PATH + "/checker.jpg", "texture_diffuse0", string.Empty));
+                textures.AddRange(Texture.GetDefaultTextures);
             }
-            using Mesh m = new(vertices.ToArray(), indices.ToArray(), textures);
-            return m;
+            ExtractBoneWeightForVertices(vertices, mesh);
+
+            return new Mesh(vertices.ToArray(), indices.ToArray(), textures);
+
         }
-        private List<Texture> LoadMaterialTexture(Material mat, TextureType type, string typeName)
+        private List<Texture> LoadMaterialTexture(Material mat, Assimp.TextureType type, Rendering.Enums.TextureType typeName)
         {
             List<Texture> textures = new();
             for (int i = 0; i < mat.GetMaterialTextureCount(type); i++)
@@ -139,7 +369,7 @@ namespace Engine.GameObjects
                         textures.Add(texture);
                         _loadedTextures.Add(texture);
                     }
-                    
+
                 }
             }
 
@@ -151,7 +381,7 @@ namespace Engine.GameObjects
             {
                 foreach (var item in m.Meshes)
                 {
-                    Meshes.Add(new Mesh(item.ObjectSetupper.GetVertices, item.ObjectSetupper.GetIndices, item.Textures, item.ObjectSetupper.GetVAOClass()));
+                    Meshes.Add(new Mesh(item.ObjectSetupper.GetVertices(), item.ObjectSetupper.GetIndices(), item.Textures, item.ObjectSetupper.GetVAOClass()));
                 }
             }
             else
@@ -165,23 +395,27 @@ namespace Engine.GameObjects
                 LogStream.IsVerboseLoggingEnabled = true;
                 var logger = new ConsoleLogStream();
                 logger.Attach();
-                Scene scene = importer.ImportFile(_path, 
-                    PostProcessSteps.FlipUVs | 
-                    PostProcessSteps.GenerateUVCoords |
-                    PostProcessSteps.OptimizeMeshes | 
-                    PostProcessSteps.SplitLargeMeshes | 
-                    PostProcessSteps.GenerateNormals  | 
-                    PostProcessSteps.RemoveRedundantMaterials | 
-                    PostProcessSteps.ImproveCacheLocality |
-                    PostProcessSteps.FindDegenerates |
-                    PostProcessSteps.FixInFacingNormals |
-                    PostProcessSteps.CalculateTangentSpace
-                    );
-                logger.Detach();
+                Scene scene = importer.ImportFile(_path, PostProcessPreset.TargetRealTimeMaximumQuality
+                    | PostProcessSteps.Debone
+                    | PostProcessSteps.FixInFacingNormals
+                    | PostProcessSteps.FlipUVs
+                    | PostProcessSteps.GenerateSmoothNormals
+                    | PostProcessSteps.GenerateUVCoords
+                    | PostProcessSteps.CalculateTangentSpace
+                    | PostProcessSteps.Triangulate);
+                Animations = new Dictionary<string, Animations.Animation>();
+
                 ProcessNode(scene.RootNode, scene);
                 _models.Add(_path, this);
+                for (int i = 0; i < scene.AnimationCount; i++)
+                {
+                    Animations.Add(scene.Animations[i].Name.ToString(), new Animations.Animation(_path, this, i));
+                }
+                Animator = new Animator(Animations.FirstOrDefault().Value);
+                logger.Detach();
             }
         }
+
         #endregion
 
         #endregion
