@@ -3,13 +3,12 @@ using Engine.GLObjects.Textures;
 using Engine.Rendering.GameObjects;
 using Engine.Structs;
 using GameEngine.Enums;
-using OpenTK;
 using OpenTK.Mathematics;
 using System.Collections.Generic;
 
 namespace Engine.Rendering.DefaultMeshes
 {
-    public static class Cube
+    public static class DefaultMesh
     {
         private static readonly Vertex[] _verttices = {
         // positions
@@ -56,12 +55,14 @@ namespace Engine.Rendering.DefaultMeshes
         new Vertex(new Vector3( 1.0f, -1.0f,  1.0f), new Vector2(0.0f, 1.0f))
     };
 
-        private static Mesh _mesh = new Mesh(_verttices);
-        public static Model GetModel() => new Model(_mesh);
-        public static Model GetTexturedModel(Texture texture)
+        private static readonly Mesh _mesh = new Mesh(_verttices, null);
+        public static Model GetCube() => new Model(_mesh, "Cube");
+        public static Model GetDefaultCube = new Model(_mesh, Texture.GetDefaultTextures, "Cube");
+        public static Model GetTexturedCube(Texture texture)
         {
             var textures = new List<Texture>() { texture };
-            return new Model(new Mesh(_verttices, textures: textures, vao: _mesh.ObjectSetupper.GetVAOClass(),setupLevel:SetupLevel.TexCoords));
+            return new Model(new Mesh(_verttices, textures: textures, vao: _mesh.ObjectSetupper.GetVAOClass(), setupLevel: SetupLevel.TexCoords), "CubeTextured");
         }
+        public static Model GetSphere = new(Game.OBJ_PATH + "Sphere.obj");
     }
 }
