@@ -8,12 +8,12 @@ namespace Engine.Rendering.DefaultMeshes
     {
         private static Vertex[] _quadVertices = {
             // positions        // texture Coords
-            new Vertex(new Vector3( -1.0f,  1.0f, 0.0f),new Vector2( 0,1)),
-            new Vertex(new Vector3(-1.0f, -1.0f, 0.0f),new Vector2( 0.0f, 0.0f)),
+            new Vertex(new Vector3( -1.0f,  1.0f, 0.0f),new Vector2(0,1)),
+            new Vertex(new Vector3(-1.0f, -1.0f, 0.0f),new Vector2(0.0f, 0.0f)),
             new Vertex(new Vector3(1.0f,  1.0f, 0.0f), new Vector2(1.0f,1.0f)),
             new Vertex(new Vector3(1.0f, -1.0f, 0.0f),new Vector2(1.0f, 0.0f)),
             };
-        private static int[] _quadIndices =
+        private static uint[] _quadIndices =
         {
               0, 1, 2, 2, 3, 0,
         };
