@@ -1,11 +1,6 @@
 ﻿using Engine.Rendering.GameObjects;
 using Engine.Structs;
 using OpenTK.Mathematics;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Engine.Rendering.DefaultMeshes
 {
@@ -18,12 +13,12 @@ namespace Engine.Rendering.DefaultMeshes
             new Vertex(new Vector3( 1.0f,  1.0f, 0.0f), new Vector2(1.0f, 1.0f)),
             new Vertex(new Vector3(1.0f, -1.0f, 0.0f),new Vector2( 1.0f, 0.0f)),
             };
-        private static readonly int[] _quadIndices =
+        private static readonly uint[] _quadIndices =
         {
             0, 1, 2, 2, 3, 0,
         };
         private static readonly Mesh _quad = new Mesh(_quadVertices, _quadIndices);
 
-        public static Mesh GetQuad() => new(_quad.ObjectSetupper.GetVertices, _quad.ObjectSetupper.GetIndices, vao: _quad.ObjectSetupper.GetVAOClass());
+        public static Mesh GetQuad() => new(_quad.ObjectSetupper.GetVertices(), _quad.ObjectSetupper.GetIndices(), vao: _quad.ObjectSetupper.GetVAOClass());
     }
 }
