@@ -1,17 +1,19 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Engine.Rendering.Enums
 {
-    [Flags] public enum RenderFlags
+    [Flags]
+    public enum RenderFlags
     {
         None = 0,
         Mesh = 1,
         Textures = 2,
         ReverseNormals = 4,
+        LightData = 8,
+        ProcessShadow = 16,
+        Animation = 32,
+        MeshAndAnimations = Mesh | Animation,
         MeshAndTextures = Mesh | Textures,
+        MeshTexturesAnimation = Mesh | Textures | Animation,
     }
 }
