@@ -3,9 +3,10 @@
     public enum SetupLevel
     {
         Position = 1,
-        TexCoords = 2,
-        Normal = 3,
+        Normal = 2,
+        TexCoords = 3,
         Tangent = 4,
-        Bitanget = 5
+        Bitanget = 5,
+        Animation = 6,
     }
 }
