@@ -1,12 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Engine.GLObjects.FrameBuffers.Interfaces
 {
-    public interface IFrameBuffer
+    public interface IFrameBuffer : IDisposable
     {
         public int FBO { get; }
         public void Setup();
