@@ -1,7 +1,7 @@
 ﻿using Engine.Components;
 using Engine.Rendering.Enums;
-using Engine.Rendering.GameObjects;
 using OpenTK.Mathematics;
+using System.Threading.Tasks;
 
 namespace Engine.GameObjects.Lights
 {
@@ -9,18 +9,18 @@ namespace Engine.GameObjects.Lights
     {
         private static int _pointLightId = -1;
         public int PointLightID;
-        public  Matrix4[] LightSpaceMatrices = new Matrix4[6];
+        public Matrix4[] LightSpaceMatrices = new Matrix4[6];
 
-        private Vector3[] _directions =
+        private readonly Vector3[] _directions =
         {
             new Vector3(1.0f,  0.0f,  0.0f ),
-            new Vector3(-1.0f,  0.0f,  0.0f ),
+            new Vector3(-1.0f, 0.0f,  0.0f ),
             new Vector3(0.0f,  1.0f,  0.0f ),
             new Vector3(0.0f, -1.0f,  0.0f ),
             new Vector3(0.0f,  0.0f,  1.0f ),
             new Vector3(0.0f,  0.0f, -1.0f )
         };
-        private Vector3[] _ups =
+        private readonly Vector3[] _ups =
         {
             new Vector3(0.0f, -1.0f,  0.0f ),
             new Vector3(0.0f, -1.0f,  0.0f ),
@@ -31,7 +31,7 @@ namespace Engine.GameObjects.Lights
         };
 
         public PointLight(Model model, LightData lightData, Transform transform)
-            : base(model, lightData,  transform)
+            : base(model, lightData, transform)
         {
             _pointLightId++;
             PointLightID = _pointLightId;
@@ -42,32 +42,39 @@ namespace Engine.GameObjects.Lights
             ShadowData.AttachCubeMap();
             FarPlane = 100.0f;
             NearPlane = 1.0f;
-            UpdateMatrices();
-
         }
-
 
         public override void Render(Shader shader, RenderFlags flags)
         {
-            shader.Use();
             string name = $"pointLights[{PointLightID}].";
-            LightData.Render(shader, name);
-            shader.SetFloat(name + "constant", 1.0f);
-            shader.SetFloat(name + "linear", 0.0014f);
-            shader.SetFloat(name + "quadratic", 0.07f);
-            shader.SetFloat(name + "farPlane", FarPlane);
-            Transform.SetValuesShader(shader, name);
-            ProcessShadow(shader);
-            base.Render(shader,flags);
+            Game.EngineGL.UseShader(shader)
+                .SetShaderData("lightColor", LightData.Color);
+            if (flags.HasFlag(RenderFlags.LightData))
+            {
+                Transform.SetLightValuesShader(shader, name);
+                LightData.Render(shader, name, flags);
+                Game.EngineGL.SetShaderData(name + "constant", 1.0f)
+                    .SetShaderData(name + "linear", 0.0014f)
+                    .SetShaderData(name + "quadratic", 0.07f)
+                    .SetShaderData(name + "farPlane", FarPlane);
+            }
+            if (flags.HasFlag(RenderFlags.ProcessShadow) && Game.Shadows)
+            {
+                ShadowData.Render(this);
+            }
+            base.Render(shader, flags);
         }
         public override Matrix4 GetProjection => Matrix4.CreatePerspectiveFieldOfView(MathHelper.DegreesToRadians(90), 1.0f, NearPlane, FarPlane);
-        public override void UpdateMatrices()
+        public override async Task UpdateMatricesAsync()
         {
             for (int i = 0; i < LightSpaceMatrices.Length; i++)
             {
-                LightSpaceMatrices[i] = Matrix4.LookAt(Transform.Position, Transform.Position + _directions[i], _ups[i]) * GetProjection;
+                await Task.Run(() =>
+                {
+                    LightSpaceMatrices[i] = Matrix4.LookAt(Transform.Position, Transform.Position + _directions[i], _ups[i]) * GetProjection;
+                });
             }
-        }
 
+        }
     }
 }
