@@ -7,20 +7,32 @@ using System.Linq;
 
 namespace Engine
 {
-    public class Shader
+    /// <summary>
+    /// Класс шейдера
+    /// </summary>
+    public class Shader : IDisposable
     {
-        public readonly int Handle;
+        /// <summary>
+        /// Индекс шейдера
+        /// </summary>
+        public readonly int ID;
 
+        /// <summary>
+        /// Возможные значения, которые нужно заполнить в шейдере
+        /// </summary>
         public readonly Dictionary<string, int> UniformLocations;
-        private string _path;
+        private readonly string _path;
+        private static int _ActiveHandle;
+        private bool disposedValue = false;
 
+        /// <summary>
+        /// Конструктор для получения шейдера с геометрическим шейдером
+        /// </summary>
+        /// <param name="vertPath">Путь к вершинному шейдеру</param>
+        /// <param name="fragPath">Путь к фрагментному шейдеру</param>
+        /// <param name="geomPath">Путь к геометрическому шейдеру</param>
         public Shader(string vertPath, string fragPath, string geomPath)
         {
-            // There are several different types of shaders, but the only two you need for basic rendering are the vertex and fragment shaders.
-            // The vertex shader is responsible for moving around vertices, and uploading that data to the fragment shader.
-            //   The vertex shader won't be too important here, but they'll be more important later.
-            // The fragment shader is responsible for then converting the vertices to "fragments", which represent all the data OpenGL needs to draw a pixel.
-            //   The fragment shader is what we'll be using the most here.
             _path = vertPath;
             // Load vertex shader and compile
             var shaderSource = File.ReadAllText(vertPath);
@@ -42,28 +54,28 @@ namespace Engine
             CompileShader(fragmentShader);
 
             shaderSource = File.ReadAllText(geomPath);
-            int geometryShader  = GL.CreateShader(ShaderType.GeometryShader);
+            int geometryShader = GL.CreateShader(ShaderType.GeometryShader);
             GL.ShaderSource(geometryShader, shaderSource);
             CompileShader(geometryShader);
 
 
             // These 3 shaders must then be merged into a shader program, which can then be used by OpenGL.
             // To do this, create a program...
-            Handle = GL.CreateProgram();
+            ID = GL.CreateProgram();
 
             // Attach  shaders...
-            GL.AttachShader(Handle, vertexShader);
-            GL.AttachShader(Handle, fragmentShader);
+            GL.AttachShader(ID, vertexShader);
+            GL.AttachShader(ID, fragmentShader);
 
-            GL.AttachShader(Handle, geometryShader);
+            GL.AttachShader(ID, geometryShader);
             // And then link them together.
-            LinkProgram(Handle);
+            LinkProgram(ID);
 
             // When the shader program is linked, it no longer needs the individual shaders attacked to it; the compiled code is copied into the shader program.
             // Detach them, and then delete them.
-            GL.DetachShader(Handle, vertexShader);
-            GL.DetachShader(Handle, fragmentShader);
-            GL.DetachShader(Handle, geometryShader);
+            GL.DetachShader(ID, vertexShader);
+            GL.DetachShader(ID, fragmentShader);
+            GL.DetachShader(ID, geometryShader);
             GL.DeleteShader(fragmentShader);
             GL.DeleteShader(vertexShader);
             GL.DeleteShader(geometryShader);
@@ -73,31 +85,32 @@ namespace Engine
             // later.
 
             // First, we have to get the number of active uniforms in the shader.
-            GL.GetProgram(Handle, GetProgramParameterName.ActiveUniforms, out var numberOfUniforms);
+            GL.GetProgram(ID, GetProgramParameterName.ActiveUniforms, out var numberOfUniforms);
 
             // Next, allocate the dictionary to hold the locations.
             UniformLocations = new Dictionary<string, int>();
 
             // Loop over all the uniforms,
-            for (var i = 0; i < numberOfUniforms ; i++)
+            for (var i = 0; i < numberOfUniforms; i++)
             {
                 // get the name of this uniform,
-                var key = GL.GetActiveUniform(Handle, i, out _, out _);
+                var key = GL.GetActiveUniform(ID, i, out _, out _);
 
                 // get the location,
-                var location = GL.GetUniformLocation(Handle, key);
+                var location = GL.GetUniformLocation(ID, key);
 
                 // and then add it to the dictionary.
                 UniformLocations.Add(key, location);
             }
         }
+
+        /// <summary>
+        /// Конструктор для получения шейдера только с вершинным и фрагментным шейдерами
+        /// </summary>
+        /// <param name="vertPath">Путь к вершинному шейдеру</param>
+        /// <param name="fragPath">Путь к фрагментному шейдеру</param>
         public Shader(string vertPath, string fragPath)
         {
-            // There are several different types of shaders, but the only two you need for basic rendering are the vertex and fragment shaders.
-            // The vertex shader is responsible for moving around vertices, and uploading that data to the fragment shader.
-            //   The vertex shader won't be too important here, but they'll be more important later.
-            // The fragment shader is responsible for then converting the vertices to "fragments", which represent all the data OpenGL needs to draw a pixel.
-            //   The fragment shader is what we'll be using the most here.
             _path = vertPath;
             // Load vertex shader and compile
             var shaderSource = File.ReadAllText(vertPath);
@@ -116,23 +129,23 @@ namespace Engine
             var fragmentShader = GL.CreateShader(ShaderType.FragmentShader);
             GL.ShaderSource(fragmentShader, shaderSource);
             CompileShader(fragmentShader);
-            Handle = GL.CreateProgram();
+            ID = GL.CreateProgram();
 
 
             // These two shaders must then be merged into a shader program, which can then be used by OpenGL.
             // To do this, create a program...
 
             // Attach both shaders...
-            GL.AttachShader(Handle, vertexShader);
-            GL.AttachShader(Handle, fragmentShader);
+            GL.AttachShader(ID, vertexShader);
+            GL.AttachShader(ID, fragmentShader);
 
             // And then link them together.
-            LinkProgram(Handle);
+            LinkProgram(ID);
 
             // When the shader program is linked, it no longer needs the individual shaders attacked to it; the compiled code is copied into the shader program.
             // Detach them, and then delete them.
-            GL.DetachShader(Handle, vertexShader);
-            GL.DetachShader(Handle, fragmentShader);
+            GL.DetachShader(ID, vertexShader);
+            GL.DetachShader(ID, fragmentShader);
             GL.DeleteShader(fragmentShader);
             GL.DeleteShader(vertexShader);
 
@@ -141,7 +154,7 @@ namespace Engine
             // later.
 
             // First, we have to get the number of active uniforms in the shader.
-            GL.GetProgram(Handle, GetProgramParameterName.ActiveUniforms, out var numberOfUniforms);
+            GL.GetProgram(ID, GetProgramParameterName.ActiveUniforms, out var numberOfUniforms);
 
             // Next, allocate the dictionary to hold the locations.
             UniformLocations = new Dictionary<string, int>();
@@ -150,16 +163,21 @@ namespace Engine
             for (var i = 0; i < numberOfUniforms; i++)
             {
                 // get the name of this uniform,
-                var key = GL.GetActiveUniform(Handle, i, out _, out _);
+                var key = GL.GetActiveUniform(ID, i, out _, out _);
 
                 // get the location,
-                var location = GL.GetUniformLocation(Handle, key);
+                var location = GL.GetUniformLocation(ID, key);
 
                 // and then add it to the dictionary.
                 UniformLocations.Add(key, location);
             }
         }
 
+        /// <summary>
+        /// Компиляция шейддера
+        /// </summary>
+        /// <param name="shader"> индекс шейдера</param>
+        /// <exception cref="Exception">Если ошибка компиляции выдает ошибку</exception>
         private static void CompileShader(int shader)
         {
             // Try to compile the shader
@@ -175,6 +193,11 @@ namespace Engine
             }
         }
 
+        /// <summary>
+        /// Присоединение разных шейдеров в одну программу
+        /// </summary>
+        /// <param name="program">индекс программы</param>
+        /// <exception cref="Exception">Выдает ошибку, если при соединении вышла ошибка, например, когда шейдеры имеют разные in - out переменные </exception>
         private void LinkProgram(int program)
         {
             // We link the program
@@ -190,84 +213,108 @@ namespace Engine
             }
         }
 
-        // A wrapper function that enables the shader program.
+        /// <summary>
+        /// Активация шейдера
+        /// </summary>
         public void Use()
         {
-            GL.UseProgram(Handle);
+            if (_ActiveHandle != ID)
+            {
+                GL.UseProgram(ID);
+                _ActiveHandle = ID;
+            }
         }
-
-        // The shader sources provided with this project use hardcoded layout(location)-s. If you want to do it dynamically,
-        // you can omit the layout(location=X) lines in the vertex shader, and use this in VertexAttribPointer instead of the hardcoded values.
-        public int GetAttribLocation(string attribName)
-        {
-            return GL.GetAttribLocation(Handle, attribName);
-        }
-
-        // Uniform setters
-        // Uniforms are variables that can be set by user code, instead of reading them from the VBO.
-        // You use VBOs for vertex-related data, and uniforms for almost everything else.
-
-        // Setting a uniform is almost always the exact same, so I'll explain it here once, instead of in every method:
-        //     1. Bind the program you want to set the uniform on
-        //     2. Get a handle to the location of the uniform with GL.GetUniformLocation.
-        //     3. Use the appropriate GL.Uniform* function to set the uniform.
 
         /// <summary>
-        /// Set a uniform int on this shader.
+        /// Получение индекс расположения по названию переменной
+        /// Например, в шейдере можно указать строго заданные места переменных, с помощью location(x)... где x - индекс расположения переменной
         /// </summary>
-        /// <param name="name">The name of the uniform</param>
-        /// <param name="data">The data to set</param>
-        
+        /// <param name="attribName">название переменной</param>
+        /// <returns></returns>
+
+        public int GetAttribLocation(string attribName)
+        {
+            return GL.GetAttribLocation(ID, attribName);
+        }
+
+        /// <summary>
+        /// Отправка в шейдер целочисленную переменную 
+        /// </summary>
+        /// <param name="name">Название переменной в шейдере</param>
+        /// <param name="data">Отправляемая переменная</param>
         public void SetInt(string name, int data)
         {
             GL.Uniform1(GetUniformLocation(name), data);
         }
+
+        /// <summary>
+        /// Отправка в шейдер булевую переменную
+        /// в GLSL нельзя вроде отправить булевую переменную поэтому перевод в целочисленный тип,
+        /// где 1 == true, а 0 == false
+        /// </summary>
+        /// <param name="name">Название переменной в шейдере</param>
+        /// <param name="data">Отправляемая переменная</param>
         public void SetInt(string name, bool data)
         {
-            int value = data? 1:0;
+            int value = data ? 1 : 0;
             GL.Uniform1(GetUniformLocation(name), value);
         }
 
         /// <summary>
-        /// Set a uniform float on this shader.
+        /// Отправка в шейдер плавающее число
         /// </summary>
-        /// <param name="name">The name of the uniform</param>
-        /// <param name="data">The data to set</param>
+        /// <param name="name">Название переменной в шейдере</param>
+        /// <param name="data">Отправляемая переменная</param>
         public void SetFloat(string name, float data)
         {
             GL.Uniform1(GetUniformLocation(name), data);
         }
 
         /// <summary>
-        /// Set a uniform Matrix4 on this shader
+        /// Отправка в шейдер матрицу 4x4
         /// </summary>
-        /// <param name="name">The name of the uniform</param>
-        /// <param name="data">The data to set</param>
-        /// <remarks>
-        ///   <para>
-        ///   The matrix is transposed before being sent to the shader.
-        ///   </para>
-        /// </remarks>
+        /// <param name="name">Название переменной в шейдере</param>
+        /// <param name="data">Отправляемая переменная</param>
         public void SetMatrix4(string name, Matrix4 data)
         {
             GL.UniformMatrix4(GetUniformLocation(name), false, ref data);
         }
 
         /// <summary>
-        /// Set a uniform Vector3 on this shader.
+        /// Отправка в шейдер 3D вектор
         /// </summary>
-        /// <param name="name">The name of the uniform</param>
-        /// <param name="data">The data to set</param>
+        /// <param name="name">Название переменной в шейдере</param>
+        /// <param name="data">Отправляемая переменная</param>
         public void SetVector3(string name, Vector3 data)
         {
             GL.Uniform3(GetUniformLocation(name), data);
-
         }
+
+        /// <summary>
+        /// Отправка в шейдер 2D вектор
+        /// </summary>
+        /// <param name="name">Название переменной в шейдере</param>
+        /// <param name="data">Отправляемая переменная</param>
         public void SetVector2(string name, Vector2 data)
         {
             GL.Uniform2(GetUniformLocation(name), data);
         }
+
+        /// <summary>
+        /// Отправка в шейдер 3D вектор
+        /// </summary>
+        /// <param name="name">Название переменной в шейдере</param>
+        /// <param name="x">координата X</param>
+        /// <param name="y">координата Y</param>
+        /// <param name="z">координата Z</param>
         public void SetVector3(string name, float x, float y, float z) => SetVector3(name, new Vector3(x, y, z));
+
+        /// <summary>
+        /// Получение расположение переменной в шейдере,
+        /// Если нам уже известно расположение,то возвращаем,
+        /// иначе ищем с помощью <see cref="GL"/>
+        /// </summary>
+        /// <param name="name">Название переменной</param>
         private int GetUniformLocation(string name)
         {
             if (UniformLocations.ContainsKey(name))
@@ -276,14 +323,42 @@ namespace Engine
             }
             else
             {
-                int pos = GL.GetUniformLocation(Handle, name);
+                int pos = GL.GetUniformLocation(ID, name);
                 UniformLocations.Add(name, pos);
 
                 if (pos == -1)
                 {
-                    Console.WriteLine($"{name} Was not setted in {_path.Split('/').Where(s=>s.Contains('.')).ToList()[3]}");
+                    Console.WriteLine($"{name} Was not setted in {_path.Split('/').Where(s => s.Contains('.')).ToList()[4]}");
                 }
                 return pos;
+            }
+        }
+
+        /// <summary>
+        /// Удаление шейдера из памяти и тд
+        /// </summary>
+        public void Dispose()
+        {
+            Dispose(true);
+            GC.SuppressFinalize(this);
+        }
+        /// <summary>
+        /// Удаление шейдера из памяти и тд
+        /// </summary>
+        ~Shader()
+        {
+            GL.DeleteProgram(ID);
+        }
+        /// <summary>
+        /// Удаление шейдера из памяти и тд
+        /// </summary>
+        protected virtual void Dispose(bool disposing)
+        {
+            if (!disposedValue)
+            {
+                GL.DeleteProgram(ID);
+
+                disposedValue = true;
             }
         }
     }
