@@ -1,17 +1,18 @@
-﻿using OpenTK;
-using OpenTK.Mathematics;
+﻿using OpenTK.Mathematics;
 
 namespace Engine.Structs
 {
     public struct Vertex
     {
-        public static int Size => (3 + 3 + 2 + 3 + 3) * sizeof(float);
-        public static int SizePositionOnly => 3 * sizeof(float);
+        public const int MAX_BONE_INFLUENCE = 4;
+        public static int Size => ((3 + 3 + 2 + 3 + 3 + 4) * sizeof(float)) + 4 * sizeof(int);
         public Vector3 Position;
         public Vector3 Normal;
         public Vector2 TexCoords;
         public Vector3 Tangent;
         public Vector3 Bitangent;
+        public Vector4i BoneIDs;
+        public Vector4 Weights;
         public Vertex(Vector3 position)
         {
             Position = position;
@@ -19,40 +20,45 @@ namespace Engine.Structs
             TexCoords = new Vector2();
             Tangent = new Vector3();
             Bitangent = new Vector3();
+            BoneIDs = new Vector4i(-1);
+            Weights = new Vector4(0);
         }
         public Vertex(Vector3 position, Vector2 texCoords)
+            : this(position)
         {
             Position = position;
-            Normal = new Vector3();
             TexCoords = texCoords;
-            Tangent = new Vector3();
-            Bitangent = new Vector3();
         }
-        public Vertex(Vector3 position,Vector3 normal, Vector2 texCoords)
+        public Vertex(Vector3 position, Vector3 normal, Vector2 texCoords)
+            : this(position, texCoords)
         {
             Position = position;
             Normal = normal;
             TexCoords = texCoords;
-            Tangent = new Vector3();
-            Bitangent = new Vector3();
         }
 
 
         public Vertex(float positionX, float positionY, float positionZ)
+            : this(new Vector3(positionX, positionY, positionZ))
         {
-            Position = new Vector3(positionX,positionY,positionZ);
-            Normal = new Vector3();
-            TexCoords = new Vector2();
-            Tangent = new Vector3();
-            Bitangent = new Vector3();
         }
         public Vertex(Vector3 position, Vector3 normal, Vector2 texCoords, Vector3 tangent, Vector3 bitangent)
+            : this(position, normal, texCoords)
         {
-            Position = position;
-            Normal = normal;
-            TexCoords = texCoords;
             Tangent = tangent;
             Bitangent = bitangent;
+        }
+
+        public Vertex(Vector3 position, Vector3 normal, Vector2 texCoords, Vector3 tangent, Vector3 bitangent, Vector4i boneIDs)
+            : this(position, normal, texCoords, tangent, bitangent)
+        {
+            BoneIDs = boneIDs;
+        }
+
+        public Vertex(Vector3 position, Vector3 normal, Vector2 texCoords, Vector3 tangent, Vector3 bitangent, Vector4i boneIDs, Vector4 weights)
+            : this(position, normal, texCoords, tangent, bitangent, boneIDs)
+        {
+            Weights = weights;
         }
     }
     public static class VertexExtensions
@@ -83,17 +89,17 @@ namespace Engine.Structs
             }
             return res;
         }
-       /* public static Vertex[] ConvertFromFloatArray(this float[] ar, bool hasNormals= false, bool hasTexCoords=false, bool hasTangets=false, bool hasBitangents=false)
-        {
-            Vertex[] result = new Vertex[ar.Length / Vertex.Size / sizeof(float)];
-            for (int i = 0; i < ar.Length - ; i++)
-            {
+        /* public static Vertex[] ConvertFromFloatArray(this float[] ar, bool hasNormals= false, bool hasTexCoords=false, bool hasTangets=false, bool hasBitangents=false)
+         {
+             Vertex[] result = new Vertex[ar.Length / Vertex.Size / sizeof(float)];
+             for (int i = 0; i < ar.Length - ; i++)
+             {
 
-                if (true)
-                {
+                 if (true)
+                 {
 
-                }
-            }
-        }*/
+                 }
+             }
+         }*/
     }
 }
